REFACTOR: Move graphs and symmetric graphs as C-sets into own module.
