+++ src/categorical_algebra/CategoricalAlgebra.jl
@@ -6,5 +6,6 @@ include("FinSets.jl")
 include("FinRelations.jl")
 include("Permutations.jl")
 include("CSets.jl")
+include("Graphs.jl")
 
 end
