+++ test/categorical_algebra/CategoricalAlgebra.jl
@@ -24,6 +24,7 @@ end
 
 @testset "CSets.jl" begin
   include("CSets.jl")
+  include("Graphs.jl")
 end
 
 end
