+++ test/categorical_algebra/Graphs.jl
@@ -1,0 +1,56 @@
+module TestGraphs
+using Test
+
+import LightGraphs
+using Catlab.CategoricalAlgebra.Graphs
+
+# Graphs
+########
+
+g = Graph()
+@test keys(g.incident) == (:src,:tgt)
+add_vertex!(g)
+add_vertices!(g, 2)
+@test nv(g) == 3
+@test ne(g) == 0
+
+add_edge!(g, 1, 2)
+add_edge!(g, 2, 3)
+@test ne(g) == 2
+@test ne(g, 1, 2) == 1
+@test has_edge(g, 1, 2)
+@test !has_edge(g, 1, 3)
+@test outneighbors(g, 2) == [3]
+@test inneighbors(g, 2) == [1]
+@test collect(all_neighbors(g, 2)) == [1,3]
+
+add_edge!(g, 1, 2)
+@test ne(g) == 3
+@test ne(g, 1, 2) == 2
+@test collect(edges(g, 1, 2)) == [1,3]
+@test outneighbors(g, 1) == [2,2]
+@test inneighbors(g, 1) == []
+
+@test LightGraphs.DiGraph(g) == LightGraphs.path_digraph(3)
+
+# Symmetric graphs
+##################
+
+g = SymmetricGraph()
+@test keys(g.incident) == (:src,)
+
+add_vertices!(g, 3)
+@test nv(g) == 3
+@test ne(g) == 0
+
+add_edge!(g, 1, 2)
+add_edge!(g, 2, 3)
+@test ne(g) == 2
+@test collect(edges(g, 1, 2)) == [1]
+@test neighbors(g, 1) == [2]
+@test neighbors(g, 2) == [1,3]
+@test neighbors(g, 3) == [2]
+
+@test LightGraphs.Graph(g) == LightGraphs.path_graph(3)
+
+end
